feat: add styling and code for program details for department admins
